requirements.txt
absl-py==2.2.2
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.2.1
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.51.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
keras==3.9.2
keyboard==0.13.5
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
mdurl==0.1.2
ml_dtypes==0.5.1
MouseInfo==0.1.3
namex==0.0.8
numpy==1.26.4
opt_einsum==3.4.0
optree==0.15.0
packaging==24.0
pandas==2.2.3
pillow==10.3.0
protobuf==5.29.4
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.2
requests==2.32.3
rich==14.0.0
setuptools==78.1.1
six==1.16.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.0.1
typing_extensions==4.13.2
tzdata==2024.2
urllib3==2.4.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2 